requirements.txt
pycolmap
usd-core
numpy
PySide6
PyOpenGL
imageio-ffmpeg